+++ app/src/main/java/com/gb/canibuyit/model/Saving.kt
@@ -6,7 +6,7 @@ class Saving(val id: Int?,
              val spendingId: Int?,
              val amount: Double,
              val created: LocalDate,
-             val target: Double) {
+             val target: Int) {
     override fun toString(): String {
         return "Saving(spendingId=$spendingId, amount=$amount, created=$created, target=$target)"
     }
